requirements.txt
Django==3.0.9
openpyxl==3.1.2
celery==5.3.4
python-dotenv==1.0.0